% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_corpus.R
\name{build_corpus}
\alias{build_corpus}
\title{Build Corpus from Text Data}
\usage{
build_corpus(df, text_col = "text")
}
\arguments{
\item{df}{A data frame containing text data.}

\item{text_col}{The name of the column containing the text to be converted into a corpus. Defaults to "text".}
}
\value{
A text corpus of class \code{Corpus} from the \code{tm} package.
}
\description{
This function takes a data frame and constructs a text corpus from a specified column.
It uses the \code{tm} package to create a corpus suitable for text mining operations.
}
\examples{
df <- data.frame(text = c("This is a test.", "Another sentence."))
corpus <- build_corpus(df)
inspect(corpus)

}
